requirements.txt
autopep8
black
isort
pylint
flake8
pycodestyle
mypy
xenon
pyyaml

glom
azure-servicebus
azure-functions